data type change (`int`=>`double`)

+++ src/Kunc.RiotGames.Api/LolMatchV5/MissionsDto.cs
@@ -8,38 +8,38 @@ namespace Kunc.RiotGames.Api.LolMatchV5;
 public class MissionsDto : BaseDto
 {
     [JsonPropertyName("playerScore0")]
-    public int PlayerScore0 { get; set; }
+    public double PlayerScore0 { get; set; }
 
     [JsonPropertyName("playerScore1")]
-    public int PlayerScore1 { get; set; }
+    public double PlayerScore1 { get; set; }
 
     [JsonPropertyName("playerScore2")]
-    public int PlayerScore2 { get; set; }
+    public double PlayerScore2 { get; set; }
 
     [JsonPropertyName("playerScore3")]
-    public int PlayerScore3 { get; set; }
+    public double PlayerScore3 { get; set; }
 
     [JsonPropertyName("playerScore4")]
-    public int PlayerScore4 { get; set; }
+    public double PlayerScore4 { get; set; }
 
     [JsonPropertyName("playerScore5")]
-    public int PlayerScore5 { get; set; }
+    public double PlayerScore5 { get; set; }
 
     [JsonPropertyName("playerScore6")]
-    public int PlayerScore6 { get; set; }
+    public double PlayerScore6 { get; set; }
 
     [JsonPropertyName("playerScore7")]
-    public int PlayerScore7 { get; set; }
+    public double PlayerScore7 { get; set; }
 
     [JsonPropertyName("playerScore8")]
-    public int PlayerScore8 { get; set; }
+    public double PlayerScore8 { get; set; }
 
     [JsonPropertyName("playerScore9")]
-    public int PlayerScore9 { get; set; }
+    public double PlayerScore9 { get; set; }
 
     [JsonPropertyName("playerScore10")]
-    public int PlayerScore10 { get; set; }
+    public double PlayerScore10 { get; set; }
 
     [JsonPropertyName("playerScore11")]
-    public int PlayerScore11 { get; set; }
+    public double PlayerScore11 { get; set; }
 }
